LOG-2319: fix fluent config format
