--- internal/pkg/generator/forwarder/suite_test.go
@@ -1,13 +1,0 @@
-package forwarder
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestForwarder(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Forwarder Generator")
-}
